Backend/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
python-dotenv
